+++ src/rt/mod.rs
@@ -57,23 +57,23 @@ mod vv;
 pub(crate) use self::vv::VersionVec;
 
 /// Maximum number of threads that can be included in a model.
-pub(crate) const MAX_THREADS: usize = 4;
+pub const MAX_THREADS: usize = 4;
 
 /// Maximum number of atomic store history to track per-cell.
 pub(crate) const MAX_ATOMIC_HISTORY: usize = 7;
 
-pub fn spawn<F>(f: F)
+pub(crate) fn spawn<F>(f: F) -> crate::rt::thread::Id
 where
     F: FnOnce() + 'static,
 {
-    execution(|execution| {
-        execution.new_thread();
-    });
+    let id = execution(|execution| execution.new_thread());
 
     Scheduler::spawn(Box::new(move || {
         f();
         thread_done();
     }));
+
+    id
 }
 
 /// Marks the current thread as blocked